requirements.txt
coverage==3.7.1
dj-database-url==0.3.0
Django==1.8.4
django-classy-tags==0.6.2
django-cms==3.2.0.dev2
django-kronos==0.8
django-reversion==1.9.3
django-sekizai==0.8.2
Django-Select2==4.3.1
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-installer==0.7.9
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
funcsigs==0.4
html5lib==0.999999
mock==1.3.0
MySQL-python==1.2.5
pbr==1.6.0
Pillow==2.9.0
python-memcached==1.57
pytz==2015.4
requests==2.7.0
six==1.9.0
tablib==0.10.0
tzlocal==1.2
wheel==0.24.0

